Encapsulates seed generation using a class
I transferred the existing code from choose.py to the new seeds module,
and encapsulated seed generation in the SeedSelector class.  By taking
as keyword arguments the number of seeds for each supported strategy
(all default to zero), this class makes it easy to generate seed sets
for hybrid selection strategies.  Also, the pagerank method was actually
calculating iterated degree, so I replaced the degree & pagerank methods
with SeedSelector._iterated_degree_seeds; this method takes a number of
generations as a parameter, and ranks nodes by degree, iterated for that
many generations.  The SeedSelector class has a class attribute which
controls the default number (two) of generations used to calculate
iterated node degree.