Alters existing metrics to use a fixed-size heap
I've modified some code in SeedSelector's actual selector methods to use
a limited-size max heap (implemented using a min-max heap) to improve
efficiency of steps from O(n log n) to O(n log k), where n is the number
of vertices in the graph, and k is the number of seeds to be selected.
Additionally, the actual min-max heap implementation is in C++, so I
wrote a simple Cython wrapper to interface between the two, as well as a
Makefile to build the project.  For some reason, this implementation
appears to produce different seed nodes than the previous commit;
testing will reveal whether this makes a difference.