Implements betweenness centrality, random seeds
I used networkx to implement the SeedSelector._between_seeds method, for
picking seeds by maximum betweenness centrality, and added support for
semi-random seed selection.  That is, by specifying the entropy keyword
argument in the SeedSelector constructor, you may control the proportion
of extra highly ranked nodes which are cached, so that when a generation
is chosen, instead of being fixed, nodes are randomly selected from the
pool of highly ranked nodes.  Also turned the metric labels into globals
in the seed module, and made some minor bugfixes (checking for unused
measures)